feat: page link blocks use links with title slug
