+++ packages/client/tiptap/extensions/pageLinkPicker/PageLinkMenu.tsx
@@ -69,7 +69,7 @@ export const PageLinkMenu = forwardRef(
               className={
                 'group flex w-full cursor-pointer items-center space-x-2 rounded-md px-3 py-2 text-sm leading-8 text-slate-700 outline-hidden hover:bg-slate-200! hover:text-slate-900 focus:bg-slate-200 data-highlighted:bg-slate-100 data-highlighted:text-slate-900'
               }
-              onClick={() => selectItem(item.title)}
+              onClick={() => selectItem(item.pageId)}
             >
               <span>{item.title}</span>
             </div>
